+++ e_metrobus/navigation/views.py
@@ -42,27 +42,6 @@ class NavigationView(TemplateView):
 
 
 class RouteView(TemplateView):
-    template_name = "navigation/route.html"
-
-    def get(self, request, *args, **kwargs):
-        context = self.get_context_data(**kwargs)
-
-        context["stations"] = widgets.StationsWidget(
-            stations.STATIONS.get_stations(), request
-        )
-        return self.render_to_response(context)
-
-    def post(self, request, *args, **kwargs):
-        def get_stations():
-            stations_raw = request.POST["stations"]
-            start, end = stations_raw.split(",")
-            return int(start[-2:]) - 1, int(end[-2:]) - 1
-
-        request.session["stations"] = get_stations()
-        return redirect("navigation:display_route")
-
-
-class RouteDropdownView(TemplateView):
     template_name = "navigation/route_dropdown.html"
 
     def get(self, request, *args, **kwargs):